refactor(input): files
