+++ src/components/TextInput/types.ts
@@ -1,0 +1,10 @@
+import { InputHTMLAttributes } from 'react';
+
+export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
+    isError?: boolean;
+    showPassword?: boolean;
+}
+
+export interface TextInputProps extends InputProps {
+    label: string;
+}
