+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './components/Theme';
 export * from './components/Modal';
 export * from './components/Button';
-export * from './components/TextInput/TextInput';
+export * from './components/TextInput';
 export * from './components/NumberInput/NumberInput';
 export * from './components/Checkbox/Checkbox';
 export * from './components/Chip/Chip';
